Refactor MySQL tests into package mysql
