+++ server/datastore/datastore_osquery_options.go
