+++ server/datastore/mysql/mysql_test.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/DATA-DOG/go-sqlmock"
 	"github.com/VividCortex/mysqlerr"
+	"github.com/fleetdm/fleet/server/kolide"
 	"github.com/go-kit/kit/log"
 	"github.com/go-sql-driver/mysql"
 	"github.com/jmoiron/sqlx"
@@ -230,4 +231,54 @@ func TestWithRetryTxxCommitError(t *testing.T) {
 	}))
 
 	require.NoError(t, mock.ExpectationsWereMet())
+}
+
+func TestAppendListOptionsToSQL(t *testing.T) {
+	sql := "SELECT * FROM app_configs"
+	opts := kolide.ListOptions{
+		OrderKey: "name",
+	}
+
+	actual := appendListOptionsToSQL(sql, opts)
+	expected := "SELECT * FROM app_configs ORDER BY name ASC LIMIT 1000000"
+	if actual != expected {
+		t.Error("Expected", expected, "Actual", actual)
+	}
+
+	sql = "SELECT * FROM app_configs"
+	opts.OrderDirection = kolide.OrderDescending
+	actual = appendListOptionsToSQL(sql, opts)
+	expected = "SELECT * FROM app_configs ORDER BY name DESC LIMIT 1000000"
+	if actual != expected {
+		t.Error("Expected", expected, "Actual", actual)
+	}
+
+	opts = kolide.ListOptions{
+		PerPage: 10,
+	}
+
+	sql = "SELECT * FROM app_configs"
+	actual = appendListOptionsToSQL(sql, opts)
+	expected = "SELECT * FROM app_configs LIMIT 10"
+	if actual != expected {
+		t.Error("Expected", expected, "Actual", actual)
+	}
+
+	sql = "SELECT * FROM app_configs"
+	opts.Page = 2
+	actual = appendListOptionsToSQL(sql, opts)
+	expected = "SELECT * FROM app_configs LIMIT 10 OFFSET 20"
+	if actual != expected {
+		t.Error("Expected", expected, "Actual", actual)
+	}
+
+	opts = kolide.ListOptions{}
+	sql = "SELECT * FROM app_configs"
+	actual = appendListOptionsToSQL(sql, opts)
+	expected = "SELECT * FROM app_configs LIMIT 1000000"
+
+	if actual != expected {
+		t.Error("Expected", expected, "Actual", actual)
+	}
+
 }
